+++ src/app/store/podcast.selector.ts
@@ -1,15 +1,10 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
 import { PodCastState } from './podcast.reducer';
-import {
-  PodCastAuthor,
-  PodCastEntry,
-  PodCastLink,
-} from '../shared/models/podcast';
+import { PodCastEntry } from '../shared/models/podcast';
 import { LoadingState, TransmissionState } from './podcast-store.model';
 import {
   PodCastDetail,
   PodCastTrackInfo,
-  Track,
 } from '../shared/models/podcast-detail';
 
 export const featureKey = 'podCast';
@@ -26,11 +21,6 @@ export const selectIsLoading = createSelector(
   (state: PodCastState): boolean => state.callState === LoadingState.LOADING
 );
 
-export const selectIsLoaded = createSelector(
-  selectPodCast,
-  (state: PodCastState): boolean => state.callState === LoadingState.LOADED
-);
-
 export const selectPodCastEntry = createSelector(
   selectPodCast,
   (state: PodCastState): PodCastEntry[] =>
@@ -42,56 +32,25 @@ export const selectPodCastEntry = createSelector(
     })
 );
 
-export const selectPodCastLink = createSelector(
-  selectPodCast,
-  (state: PodCastState): PodCastLink[] => state.data.link
-);
-
-export const selectPodCastAuthor = createSelector(
-  selectPodCast,
-  (state: PodCastState): PodCastAuthor => state.data.author
-);
-
-export const selectPodCastId = createSelector(
-  selectPodCast,
-  (state: PodCastState): string => state.data.id.label
-);
-
-export const selectPodCastUpdated = createSelector(
-  selectPodCast,
-  (state: PodCastState): string => state.data.updated.label
-);
-
-export const selectPodCastRights = createSelector(
-  selectPodCast,
-  (state: PodCastState): string => state.data.rights.label
-);
-
-export const selectPodCastTitle = createSelector(
-  selectPodCast,
-  (state: PodCastState): string => state.data.title.label
-);
-
-export const selectPodCastIcon = createSelector(
-  selectPodCast,
-  (state: PodCastState): string => state.data.icon.label
-);
-
 export const selectCurrentPodCast = createSelector(
   selectPodCast,
-  (state: PodCastState): PodCastDetail => state.currentPodCast
+  (state: PodCastState): PodCastDetail => {
+    const entry = state.data.entry.find(
+      (e) => e.id.attributes['im:id'] === `${state.currentPodCast.info.trackId}`
+    );
+    return {
+      ...state.currentPodCast,
+      info: {
+        ...state.currentPodCast.info,
+        summary: entry?.summary.label,
+      },
+    };
+  }
 );
 
 export const selectCurrentPodCastId = createSelector(
   selectPodCast,
   (state: PodCastState): number => state.currentPodCast.info.trackId
-);
-
-export const selectCurrentEpisode = createSelector(
-  selectPodCast,
-  (state: PodCastState): Track => {
-    return state.currentTrack;
-  }
 );
 
 export const selectCurrentPodCastTrackInfo = createSelector(