+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Style improvenments
+- Params and selectors for podcast detail and episode detail
 
 ## [0.0.1] - 2023-10-25